move zlib/loading to resources/ since exe binary related
