+++ ENIGMAsystem/SHELL/Universal_System/Resources/loading.cpp
@@ -16,11 +16,11 @@
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
 
-#include "Resources/resinit.h"
-#include "Resources/sprites_internal.h"
-#include "Resources/background_internal.h"
-#include "roomsystem.h"
-#include "Object_Tiers/object.h"
+#include "resinit.h"
+#include "sprites_internal.h"
+#include "background_internal.h"
+#include "Universal_System/roomsystem.h"
+#include "Universal_System/Object_Tiers/object.h"
 #include "libEGMstd.h"
 //#include "mathnc.h"
